[ fix #9 ] Enable renaming through a rule definition
Remove the `PsiNameIdentifierOwner` from the `PestIdentifierMixin`.
Adapt the grammar so that the rule name is not also an identifier that references this rule.
For renaming we only need
`setName` on the grammar rule and
`handleElementRename` on the the identifiers that reference the rule.

Enable more recovery in the grammar.

Allow COMMENT and WHITESPACE for renaming because these rules can be overwritten.